- Cleaning code: Catalog & Helpcard all process, updated room processes, deleted redundant code;

+++ app/src/main/java/space/lopatkin/spb/helpboardgamecard/data/local/room/RoomBoardgameLocalDataSource.kt
@@ -4,7 +4,10 @@ import android.content.Context
 import kotlinx.coroutines.CoroutineScope
 import space.lopatkin.spb.helpboardgamecard.data.local.data.source.BoardgameLocalDataSource
 import space.lopatkin.spb.helpboardgamecard.data.local.room.RoomDb.Companion.getInstance
-import space.lopatkin.spb.helpboardgamecard.domain.model.*
+import space.lopatkin.spb.helpboardgamecard.domain.model.BoardgameInfo
+import space.lopatkin.spb.helpboardgamecard.domain.model.BoardgameRaw
+import space.lopatkin.spb.helpboardgamecard.domain.model.Helpcard
+import space.lopatkin.spb.helpboardgamecard.domain.model.Message
 
 class RoomBoardgameLocalDataSource(
     private val context: Context,
@@ -19,7 +22,8 @@ class RoomBoardgameLocalDataSource(
 
     override suspend fun getAllBoardgamesInfo(): Result<List<BoardgameInfo>> {
         return try {
-            Result.success(boardgameDao.getAllBoardgamesInfo())
+            val list: List<BoardgameInfo> = boardgameDao.getAllBoardgamesInfo()
+            Result.success(list)
         } catch (cause: Throwable) {
             Result.failure(cause)
         }
@@ -27,8 +31,12 @@ class RoomBoardgameLocalDataSource(
 
     override suspend fun getHelpcardBy(boardgameId: Long): Result<Helpcard> {
         return try {
-            val data: Helpcard = boardgameDao.getHelpcardBy(boardgameId)
-            Result.success(data)
+            val dataFromDb: Helpcard? = boardgameDao.getHelpcardBy(boardgameId)
+            if (dataFromDb != null) {
+                Result.success(dataFromDb)
+            } else {
+                Result.failure(Exception("NotFoundException (room): query (getHelpcardByBoardgameId) not finished because Helpcard not found in db table"))
+            }
         } catch (cause: Throwable) {
             Result.failure(cause)
         }
@@ -67,12 +75,12 @@ class RoomBoardgameLocalDataSource(
 
     override suspend fun deleteBoardgameBy(boardgameId: Long): Result<Message> {
         return try {
-            val deletedInfo: Int = boardgameDao.deleteBoardgameInfoBy(boardgameId)
             val deletedHelpcard: Int = boardgameDao.deleteHelpcardBy(boardgameId)
+            val deletedInfo: Int = boardgameDao.deleteBoardgameInfoBy(boardgameId)
             if (deletedInfo > 0 && deletedHelpcard > 0) {
                 return Result.success(Message.DELETE_ITEM_ACTION_ENDED_SUCCESS)
             } else {
-                return Result.failure(DataPassError("Data pass is null", IllegalStateException()))
+                Result.failure(Exception("InvalidOperationException (room): models (Helpcard & BoardgameInfo) not deleted full"))
             }
         } catch (cause: Throwable) {
             Result.failure(cause)
@@ -83,9 +91,9 @@ class RoomBoardgameLocalDataSource(
         return try {
             val updated: Int = boardgameDao.update(boardgameInfo)
             if (updated > 0) {
-                return Result.success(Message.ACTION_ENDED_SUCCESS)
+                Result.success(Message.ACTION_ENDED_SUCCESS)
             } else {
-                return Result.failure(DataPassError("Data pass is null", IllegalStateException()))
+                Result.failure(Exception("NotFoundException (room): model (BoardgameInfo) not updated because (BoardgameId) not found in db table"))
             }
         } catch (cause: Throwable) {
             Result.failure(cause)
@@ -121,26 +129,15 @@ class RoomBoardgameLocalDataSource(
     override suspend fun deleteUnlockBoardgames(): Result<Message> {
         return try {
             val listIdUnlock: Array<Long> = boardgameDao.getBoardgameIdsByUnlock()
-
-            for (index in 0 until listIdUnlock.size) {
-                val deletedStatusInfo: Int = boardgameDao.deleteBoardgameInfoBy(listIdUnlock.get(index))
-                val deletedStatusHelpcard: Int = boardgameDao.deleteHelpcardBy(listIdUnlock.get(index))
-
-                if (deletedStatusInfo == 0 && deletedStatusHelpcard == 0) {
-                    return Result.failure(DataPassError("Data pass is null", IllegalStateException()))
+            if (listIdUnlock.size > 0) {
+                for (index in 0 until listIdUnlock.size) {
+                    val deletedStatusInfo: Int = boardgameDao.deleteBoardgameInfoBy(listIdUnlock.get(index))
+                    val deletedStatusHelpcard: Int = boardgameDao.deleteHelpcardBy(listIdUnlock.get(index))
                 }
+                Result.success(Message.DELETE_ALL_ACTION_ENDED_SUCCESS)
+            } else {
+                Result.success(Message.DELETE_ALL_ACTION_STOPPED)
             }
-            Result.success(Message.DELETE_ALL_ACTION_ENDED_SUCCESS)
-        } catch (cause: Throwable) {
-            Result.failure(cause)
-        }
-    }
-
-    override suspend fun getBoardgamesIdByUnlock(): Result<Array<Long>> {
-        return try {
-            val listIdUnlock: Array<Long> = boardgameDao.getBoardgameIdsByUnlock()
-
-            Result.success(listIdUnlock)
         } catch (cause: Throwable) {
             Result.failure(cause)
         }
